Changes were made to UndoRedoManager.java to fix code smells and attempt to make it live up to at least some of the SOLID principles. Furthermore, tests unit tests and dependencies for these were implemented to test some of the changes. Resource files were moved to relevant folder for tests.

+++ jhotdraw-samples/jhotdraw-samples-misc/src/main/java/org/jhotdraw/samples/svg/gui/ViewToolBar.java
@@ -185,7 +185,7 @@ public class ViewToolBar extends AbstractToolBar {
                     public void propertyChange(PropertyChangeEvent evt) {
                         if (evt.getPropertyName() == DrawingView.SCALE_FACTOR_PROPERTY) {
                             if (evt.getNewValue() != null) {
-                                scaleFactorField.setValue((Double) evt.getNewValue());
+                                scaleFactorField.setValue(evt.getNewValue());
                             }
                         }
                     }
